+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-* Fixed missing data files
+* Fixed unable to respawn